Added Repository for the LaserDbContext
Big clearance

+++ NewLaserProject/ViewModels/LaserDbViewModel.cs
@@ -4,15 +4,27 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
+using System.Reactive.Linq;
 using System.Threading.Tasks;
 using HandyControl.Controls;
 using HandyControl.Tools.Extension;
 using MachineClassLibrary.Laser.Parameters;
+using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Toolkit.Mvvm.Input;
+using NewLaserProject.Classes;
 using NewLaserProject.Classes.ProgBlocks;
 using NewLaserProject.Data.Models;
 using NewLaserProject.Data.Models.DTOs;
+using NewLaserProject.Data.Models.MaterialEntRuleFeatures.Create;
+using NewLaserProject.Data.Models.MaterialEntRuleFeatures.Get;
+using NewLaserProject.Data.Models.MaterialEntRuleFeatures.Update;
+using NewLaserProject.Data.Models.MaterialFeatures.Create;
+using NewLaserProject.Data.Models.MaterialFeatures.Delete;
+using NewLaserProject.Data.Models.MaterialFeatures.Get;
+using NewLaserProject.Data.Models.TechnologyFeatures.Create;
+using NewLaserProject.Data.Models.TechnologyFeatures.Delete;
+using NewLaserProject.Data.Models.TechnologyFeatures.Update;
 using NewLaserProject.ViewModels.DbVM;
 using NewLaserProject.ViewModels.DialogVM;
 using NewLaserProject.Views.DbViews;
@@ -32,32 +44,46 @@ namespace NewLaserProject.ViewModels
         {
             get; set;
         }
-        public LaserDbViewModel(DbContext db, ExtendedParams defaultParams)
+        public LaserDbViewModel(DbContext db, IMediator mediator, ExtendedParams defaultParams)
         {
             _db = db;
+            _mediator = mediator;
             _defaultParams = defaultParams;
-            _db.Set<Technology>()
-                .Include(t => t.Material)
-                .LoadAsync()
+            _mediator.Send(new GetFullMaterialRequest())
                 .ContinueWith(t =>
                 {
-                    if (t.IsCompletedSuccessfully)
-                    {
-                        Technologies = _db.Set<Technology>()
-                            .Local
-                            .ToObservableCollection();
+                    var materialsRequest = t.Result;
+                    Materials = materialsRequest.Materials.ToObservableCollection();
+                    ReviseTechnologies();
+                }, TaskScheduler.Default)
+                .ConfigureAwait(false);
+                
+            //_db.Set<Technology>()
+            //    .Include(t => t.Material)
+            //    .LoadAsync()
+            //    .ContinueWith(t =>
+            //    {
+            //        if (t.IsCompletedSuccessfully)
+            //        {
+            //            Technologies = _db.Set<Technology>()
+            //                .Local
+            //                .ToObservableCollection();
 
-                        _db.Set<Material>()
-                            .Load();
+            //            //_db.Set<Material>()
+            //            //    .Load();
 
-                        Materials = _db.Set<Material>()
-                            .Local
-                            .ToObservableCollection();
-                    }
-                });
+            //            //Materials = _db.Set<Material>()
+            //            //    .Local
+            //            //    .ToObservableCollection();
+            //        }
+            //    });
         }
         private readonly DbContext _db;
+        private readonly IMediator _mediator;
         private readonly ExtendedParams _defaultParams;
+
+        private void ReviseTechnologies() => Technologies = Materials.SelectMany(m => m.Technologies).ToObservableCollection();
+
 
         [ICommand]
         private async Task AddMaterial()
@@ -70,8 +96,12 @@ namespace NewLaserProject.ViewModels
             if (result.Success)
             {
                 var newMaterial = new Material(result.CommonResult);
-                _db.Add(newMaterial);
-                _db.SaveChanges();
+
+                var request = new CreateGetMaterialRequest(newMaterial);
+                var response = await _mediator.Send(request);
+                Materials = response.Materials.ToObservableCollection();
+                //_db.Add(newMaterial);
+                //_db.SaveChanges();
             }
         }
 
@@ -114,46 +144,49 @@ namespace NewLaserProject.ViewModels
                 newTechnology.ProcessingProgram = writeTechVM.TechnologyWizard.SaveListingToFolder(path);
 
                 newTechnology.ProgramName = writeTechVM.TechnologyName ?? DateTime.Now.ToString();//TODO if name isn't typed
-                _db.Add(newTechnology);
-                _db.SaveChanges();
+                //_db.Add(newTechnology);
+                //_db.SaveChanges();
+                var response = await _mediator.Send(new CreateTechnologyRequest(newTechnology));
+                ReviseTechnologies();
             }
         }
 
         [ICommand]
         private async void AssignRule(Material material)
         {
-            var matEntRule = _db.Set<MaterialEntRule>()
-                    .AsNoTracking()
-                    .SingleOrDefault(mer => mer.Material.Id == material.Id);
+            //var matEntRule = _db.Set<MaterialEntRule>()
+            //        .AsNoTracking()
+            //        .SingleOrDefault(mer => mer.Material.Id == material.Id);
 
-            var newRule = matEntRule ?? new MaterialEntRule
-            {
-                Material = material,
-            };
+            var response = await _mediator.Send(new GetRuleByMaterialIdRequest(material.Id));
+            var matEntRule = response.MaterialEntRule;
 
             var result = await Dialog.Show<CommonDialog>()
                 .SetDialogTitle("Правило обработки")
-                .SetDataContext<MaterialEntRuleVM>(vm => vm.MaterialEntRule = newRule)
+                .SetDataContext<MaterialEntRuleVM>(vm => vm.MaterialEntRule = matEntRule ?? new MaterialEntRule
+                {
+                    Material = material,
+                })
                 .GetCommonResultAsync<MaterialEntRule>();
 
             if (result.Success)
             {
                 if (matEntRule is not null)
                 {
-                    //_db.Set<MaterialEntRule>()
-                    //    .Update(matEntRule);
+                    //var rule = _db.Set<MaterialEntRule>()
+                    //.SingleOrDefault(mer => mer.Material.Id == material.Id);
+                    //rule.Offset = matEntRule.Offset;
+                    //rule.Width = matEntRule.Width;
 
-                    var rule = _db.Set<MaterialEntRule>()
-                    .SingleOrDefault(mer => mer.Material.Id == material.Id);
-                    rule.Offset = matEntRule.Offset;
-                    rule.Width = matEntRule.Width;
+                    await _mediator.Send(new UpdateMaterialEntRuleRequest(result.CommonResult)).ConfigureAwait(false);
                 }
                 else
                 {
-                    _db.Set<MaterialEntRule>()
-                        .Add(newRule);
+                    //_db.Set<MaterialEntRule>()
+                    //    .Add(result.CommonResult);
+                    await _mediator.Send(new CreateMaterialEntRuleRequest(result.CommonResult)).ConfigureAwait(false);
                 }
-                _db.SaveChanges();
+                //_db.SaveChanges();
             }
         }
 
@@ -181,8 +214,9 @@ namespace NewLaserProject.ViewModels
                     return;
                 }
                 technology.ProcessingProgram = result.CommonResult.SaveListingToFolder(ProjectPath.GetFolderPath(ProjectFolders.TECHNOLOGY_FILES));
-                _db.Set<Technology>().Update(technology);
-                _db.SaveChanges();
+                //_db.Set<Technology>().Update(technology);
+                //_db.SaveChanges();
+                var response = await _mediator.Send(new UpdateTechnologyRequest(technology));
                 File.Delete(path);
             }
         }
@@ -198,12 +232,14 @@ namespace NewLaserProject.ViewModels
             }.ShowDialog();
         }
         [ICommand]
-        private void DeleteTechnology(Technology technology)
+        private async void DeleteTechnology(Technology technology)
         {
-            _db.Set<Technology>()
-                .Remove(technology);
-            DeleteTechnologyFile(technology);
-            _db.SaveChanges();
+            //_db.Set<Technology>()
+            //    .Remove(technology);
+            var response = await _mediator.Send(new DeleteTechnologyRequest(technology));
+            if(response.IsDeleted) DeleteTechnologyFile(technology);
+            ReviseTechnologies();
+            //_db.SaveChanges();
         }
         private void DeleteTechnologyFile(Technology technology)
         {
@@ -211,17 +247,22 @@ namespace NewLaserProject.ViewModels
             File.Delete(path);
         }
         [ICommand]
-        private void DeleteMaterial(Material material)
+        private async Task DeleteMaterial(Material material)
         {
-            var deletingMaterial = _db.Remove(material);
-            if (deletingMaterial is not null)
-            {
-                deletingMaterial
-                    .Entity?
-                    .Technologies?
-                    .ForEach(t => DeleteTechnologyFile(t));
-            }
-            _db.SaveChanges();
+            var request = new DeleteGetMaterialRequest(material);
+            var response = await _mediator.Send(request);
+            Materials = response.Materials.ToObservableCollection();
+            Technologies = Materials.SelectMany(m=>m.Technologies).ToObservableCollection();
+            material.Technologies?.ForEach(DeleteTechnologyFile);
+            //var deletingMaterial = _db.Remove(material);
+            //if (deletingMaterial is not null)
+            //{
+            //    deletingMaterial
+            //        .Entity?
+            //        .Technologies?
+            //        .ForEach(t => DeleteTechnologyFile(t));
+            //}
+           // _db.SaveChanges();
         }
     }
 }
